+++ shared/src/simulation/events/index.ts
@@ -11,6 +11,7 @@ export * from './vehicles-sent';
 export * from './try-to-distribute';
 export * from './vehicle-transfer-successful';
 export * from './transfer-connection-missing';
+export * from './leader-changed';
 export * from './material-removed';
 export * from './personnel-removed';
 export * from './patient-removed';