+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 -   Simulated regions display their names on the map.
 -   There is now a button in the patients tab of a simulated region that removes the patient from the scenario
 -   There is now a button in the patients tab of a simulated region that removes the patient from the simulated region and places the patient on the map next to the simulated region (on the right hand side)
+-   Vehicles now have a property to indicate their current occupation.
 
 ## [0.4.1] - 2023-04-20